requirements.txt
discord.py==2.1.1
pytz
python-dotenv
openai==0.28.0
pytesseract
pandas
gspread
oauth2client
aiohttp
gspread
oauth2client
yt-dlp
PyNaCl